up and down command will redirect user to the dashboard for registering/deregistering

+++ cmd/down.go
@@ -1,19 +1,22 @@
 package main
 
 import (
-	"bufio"
 	"context"
 	"fmt"
+	"log"
 	"math"
 	"math/big"
-	"os"
 	"os/exec"
+	"path"
+	"runtime"
 	"strings"
 	"time"
 
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum/ethclient"
+	dnr "github.com/republicprotocol/darknode-cli/bindings"
 	"github.com/republicprotocol/republic-go/cmd/darknode/config"
 	"github.com/republicprotocol/republic-go/contract"
 	"github.com/republicprotocol/republic-go/contract/bindings"
@@ -23,7 +26,6 @@ import (
 // destroyNode tears down the deployed darknode by its name.
 func destroyNode(ctx *cli.Context) error {
 	name := ctx.Args().First()
-	force := ctx.Bool("force")
 
 	if name == "" {
 		cli.ShowCommandHelp(ctx, "down")
@@ -31,30 +33,72 @@ func destroyNode(ctx *cli.Context) error {
 	}
 
 	nodeDirectory := nodeDirectory(name)
-	if !force {
-		ip, err := getIp(nodeDirectory)
-		if err != nil {
-			return ErrNoDeploymentFound
-		}
+	ip, err := getIp(nodeDirectory)
+	if err != nil {
+		return ErrNoDeploymentFound
+	}
 
-		for {
-			fmt.Printf("You need to %sderegister your Darknode%s and %swithdraw all fees%s at\n", RED, RESET, RED, RESET)
-			fmt.Printf("https://darknode.republicprotocol.com/status/%v\n", ip)
-			fmt.Println("Have you deregistered your Darknode and withdrawn all fees? (Yes/No)")
+	config, err := config.NewConfigFromJSONFile(path.Join(nodeDirectory, "config.json"))
+	if err != nil {
+		return err
+	}
+	id := config.Address.ID()
+	network := config.Ethereum.Network
+	dnrAddress := common.HexToAddress(dnrAddress(network))
+	testnet := ethereumTestnet(network)
 
-			reader := bufio.NewReader(os.Stdin)
-			text, _ := reader.ReadString('\n')
-			input := strings.ToLower(strings.TrimSpace(text))
-			if input == "yes" || input == "y" {
-				break
-			}
-			if input == "no" || input == "n" {
-				return nil
-			}
+	// Query registry smart contract on Ethereum if the darknode has been
+	// fully de-registered
+	client, err := ethclient.Dial(fmt.Sprintf("https://%v.infura.io", testnet))
+	if err != nil {
+		log.Fatal(err)
+	}
+	registry, err := dnr.NewBindings(dnrAddress, client)
+	if err != nil {
+		return err
+	}
+	registered, err := registry.IsRegistered(&bind.CallOpts{}, common.BytesToAddress(id))
+	if err != nil {
+		return err
+	}
+
+	// Redirect the user to the de-registering URL if darknode is still registered.
+	if registered {
+		fmt.Printf("%sYour node hasn't been deregistered%s\n", RED, RESET)
+		fmt.Printf("%sYou will be redirected to deregister your node%s\n", RED, RESET)
+		time.Sleep(3 * time.Second)
+
+		var redirect *exec.Cmd
+		switch runtime.GOOS {
+		case "darwin":
+			redirect = exec.Command("open", fmt.Sprintf("https://darknode.republicprotocol.com/status/%v", ip))
+		case "linux":
+			redirect = exec.Command("xdg-open", fmt.Sprintf("https://darknode.republicprotocol.com/status/%v", ip))
+		default:
+			log.Fatal("unsupported operating system")
 		}
+		pipeToStd(redirect)
+		if err := redirect.Start(); err != nil {
+			return err
+		}
+		return redirect.Wait()
+	}
+
+	// Check if the darknode is in pending deregistration state.
+	pendingDeregistration, err := registry.IsPendingDeregistration(&bind.CallOpts{}, common.BytesToAddress(id))
+	if err != nil {
+		return err
+	}
+	if pendingDeregistration {
+		fmt.Printf("%sYour node is pending for deregistration%s\n", RED, RESET)
+		fmt.Printf("%sDarknode can only be destroyed when fully deregistred%s\n", RED, RESET)
+		fmt.Printf("%sPlease wait for it to be fully deregistered and try again%s\n", RED, RESET)
+		return nil
 	}
 
 	fmt.Printf("%sDestroying your darknode ...%s\n", GREEN, RESET)
+	time.Sleep(time.Hour)
+
 	cmd := fmt.Sprintf("cd %v && terraform destroy --force && find . -type f -not -name 'config.json' -delete", nodeDirectory)
 	destroy := exec.Command("bash", "-c", cmd)
 	pipeToStd(destroy)
@@ -196,6 +240,30 @@ func renAddress(network contract.Network) string {
 		return "0x81793734c6Cf6961B5D0D2d8a30dD7DF1E1803f1"
 	case "testnet":
 		return "0x6f429121a3bd3e6c1c17edbc676eec44cf117faf"
+	default:
+		return ""
+	}
+}
+
+// darknode registry on different testnet
+func dnrAddress(network contract.Network) string {
+	switch network {
+	case "mainnet":
+		return "0x3799006a87FDE3CCFC7666B3E6553B03ED341c2F"
+	case "testnet":
+		return "0x75Fa8349fc9C7C640A4e9F1A1496fBB95D2Dc3d5"
+	default:
+		return ""
+	}
+}
+
+// ethereumTestnet returns the testnet name of different network
+func ethereumTestnet(network contract.Network) string {
+	switch network {
+	case "mainnet":
+		return "mainnet"
+	case "testnet":
+		return "kovan"
 	default:
 		return ""
 	}
